--refactor ollamachat
