+++ Dockerfile
@@ -6,18 +6,16 @@ RUN apt-get update && apt-get install -y \
     && rm -rf /var/lib/apt/lists/*
 
 RUN curl https://ollama.com/install.sh | sh
-RUN ollama pull llama3.2
 
 WORKDIR /app
-
-RUN python -m venv /opt/venv
-ENV PATH="/opt/venv/bin:$PATH"
 
 COPY requirements.txt .
 COPY app.py .
 COPY entrypoint.sh .
 
-RUN pip install --no-cache-dir -r requirements.txt
+RUN pip install --no-cache-dir \
+    ollama \
+    requests
 
 RUN chmod +x entrypoint.sh
